// frontend/src/types.ts
/*
 Generated by typeshare 1.13.3
*/

export interface AnswerQuestionRequest {
	answer_text: string;
}

export interface AskQuestionRequest {
	question_text: string;
}

export interface QAPair {
	question_id: string;
	question_text: string;
	answer_text: string;
	answered_at: string;
}

export interface Profile {
	user_id: string;
	qa_pairs: QAPair[];
}

export interface FeedResponse {
	profiles: Profile[];
}

export interface Match {
	user_id: string;
	qa_pairs: QAPair[];
}

export interface MatchInfo {
	email: string;
}

export interface MatchesResponse {
	matches: Match[];
}

export interface MyProfileResponse {
	qa_pairs: QAPair[];
}

export interface Question {
	question_id: string;
	question_text: string;
	created_at: string;
}

export interface QuestionInboxResponse {
	questions: Question[];
}

export interface RequestMagicLinkRequest {
	email: string;
}

export interface RequestMagicLinkResponse {
	success: boolean;
}

export interface ScheduleMeetupRequest {
	match_user_id: string;
}

export interface ScheduleMeetupResponse {
	success: boolean;
}

export interface SuccessResponse {
	success: boolean;
}

export interface VerifySessionRequest {
	session_id: string;
}

export interface VerifySessionResponse {
	user_id: string;
}